:bug: Fix missed type updates to WageEnhancement model.

+++ api/src/models/wage-enhancement.ts
@@ -53,11 +53,11 @@ export class WageEnhancement extends Model<
   @NotNull
   declare employeeName: string
 
-  @Attribute(DataTypes.FLOAT)
+  @Attribute(DataTypes.DECIMAL(10, 2))
   @NotNull
   declare hoursEstimated: string
 
-  @Attribute(DataTypes.FLOAT)
+  @Attribute(DataTypes.DECIMAL(10, 2))
   @NotNull
   declare hoursActual: string
